Update issue 4048
Turns out that I completely misunderstood the Collapse Blocked Elements function from FireFox. I thought it would collapse parent element(s) if they were empty. However, it turns out that it only hides the element itself if enabled...

I removed the code for this function, enabled it by default (used 'switch' instead of another 'if' as it all checks the same variable) and now the function only hides the element. I didn't choose for .remove() as this could cause JavaScript errors on the page (if it refers to an no-longer-existing element).

Note to Michael:
.hide() sets the style to display:none, which is easily overridden by other page css rules. Therefore I used the .css() and the attr(width|height)

Going to do some improvements, so not yet 'ReadyForRelease'