Bugfix and refactor for removing custom filters
Now both popup and background call a function
in background.js to both confirm that the user
wants to remove their filters on a domain and to
actually do it. Both the confirmation and the
reloading of the tab were buggy for me unless
performed by background.js

Note that we're now prompting the user even if
there is only 1 filter on this domain. This is
because the prompt gives us the chance to tell
them the removal applies specifically to
foo.bar.com, so it's clearer what's happening.
The translation message change is for the same
reason.