Toward a fix for issue #29. Blocked by Safari team
As of Safari 6.1, tabs opened from the Top Sites
page appear with ads. This is because they are
preloaded in invisible pages, a new feature.
A tab may have one or more invisible pages.
Unfortunately, we're currently unable to block
ads on those pages before they finish loading.

In order to block ads on a page/frame, we:
- Have the content script send a message to the
background page, asking what its behavior should
be on the page or frame (block or not block)
- Have the background page send a message back
to tell it what to do

The only way to send those messages between
execution contexts is to use the API the browser
provides. Chrome has one, Safari has another, and
port.js makes Safari look like Chrome.

Two things broke in Safari 6.1.

1) The `sender` object supplied by Safari when
delivering messages no longer supplied the
sending tab's url. We need to know this, because
the blocking rules for a frame on SiteA which
points to SiteB depend on both of those urls.

We were able to fix this by adding the function
safari_get_tab_url.for_frame and passing extra
info to BGcall('get_content_script_data')

2) The messageEvent object used by dispatchContext
no longer provides .target.page to tell us what
page we should send back a "hide or don't hide"
message to. There doesn't seem to be any way to
get access to such a page object.

This part we can't resolve, as far as I know.
If a tab has several (possibly invisible) pages,
yet the SafariBrowserTab object has a .page
property that's undefined, we have handle to send
a message to that page.

We've notified the Safari team of this issue and
will have to wait on their response.