Handle prerendering of Top Sites in Safari 6.1.
Fixes #29.

Background:

As of Safari 6.1, tabs opened from the Top Sites
page appear with ads. This is because they are
preloaded in invisible pages, a new feature.
A tab may have one or more invisible pages.
Unfortunately, we're currently unable to block
ads on those pages before they finish loading.

In order to block ads on a page/frame, we:
- Have the content script send a message to the
background page, asking what its behavior should
be on the page or frame (block or not block)
- Have the background page send a message back
to tell it what to do

The only way to send those messages between
execution contexts is to use the API the browser
provides. Chrome has one, Safari has another, and
port.js makes Safari look like Chrome.

Two things broke in Safari 6.1.

1) The messageEvent.target object supplied by
Safari when delivering messages no longer supplied
the sending tab's url. We need to know this,
because the blocking rules for a frame on SiteA
which points to SiteB depend on both of those
urls.  However, now that the sending tab may be
rendering a regular top-level page or an invisible
prerendered top-level page, we would need to know
a per-page-per-tab .url rather than a per-tab
.url.

We were able to fix this by adding the
chrome._tabInfo helper object in port.js, and
recording the URL for every visible and invisible
page in every tab whenever a top-level frame made
a request.  This data is used in canLoad and in
chrome.extension.onRequest to determine the .id
and .url of the requesting page.

2) The messageEvent object used by dispatchContext
no longer provides .target.page to tell us what
page we should send back a "hide or don't hide"
message to. There is no way to get access to such
a page object.  This has been fixed in Safari
6.1.1 and Safari 7.0.1 after discussion with the
Safari team.

Note that our current solution assumes a maximum
of one invisible page per tab, which Safari team
said we could assume for now.  If a future version
of Safari has several invisible pages per tab
(perhaps prerendering several links), our scheme
will fail and we'll likely need an API change from
Safari to support figuring out the requesting
page's URL.