Clarify license and revert last change.
We also copy the comment at the top of youtube_safari_only, so it's not
sufficient to give credit only down low.  I moved the notice to the top
of the file.

I also reverted Tomas's last changes, as his comment didn't explain their
point... one of them changes something we copied from working ABP code, and
one turns log() into console.log().  If the change to the ABP code is
necessary, explain more clearly in a comment.  I just verified that log()
works fine, and that's what we use in place of console.log almost
everywhere else in the code so we don't print needless debug messages.