Missing semis after var declarations.
Oops, at least one of these missing semis was mine: validTab's
declaration never ended in a semicolon.

Just like for
  var x = 4;
we should put semis after
  var x = function() {...};

I actually don't recall the risk to *not* putting the semis in the
latter case, but IIRC, "JavaScript: The Good Parts" recommends not
relying on JS's optional semis.