Make right-click-opening "here" link work.
If you right-click the "here" link on the update page and choose
Open in New Tab, the code would just open the update page again.
Use <a target="_blank"> instead.

Note that if the user does the same to the "disable this" link or the
"Want more details?" link, they're out of luck, but that's OK.